server/pyproject.toml
[project]
name = "kalender-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "pillow-heif>=1.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]
